+++ src/components/manage/UniversalLink/UniversalLink.test.jsx
@@ -5,6 +5,7 @@ import configureStore from 'redux-mock-store';
 import { render } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import UniversalLink from './UniversalLink';
+import config from '@plone/volto/registry';
 
 const mockStore = configureStore();
 const store = mockStore({
@@ -119,6 +120,27 @@ describe('UniversalLink', () => {
 
     expect(getByTitle('Volto GitHub repository').getAttribute('target')).toBe(
       null,
+    );
+  });
+
+  it('check UniversalLink renders ext link for blacklisted urls', () => {
+    config.settings.internalUrlBlacklist = ['/external-app'];
+
+    const { getByTitle } = render(
+      <Provider store={store}>
+        <MemoryRouter>
+          <UniversalLink
+            href="http://localhost:3000/external-app"
+            title="Blacklisted route"
+          >
+            <h1>Title</h1>
+          </UniversalLink>
+        </MemoryRouter>
+      </Provider>,
+    );
+
+    expect(getByTitle('Blacklisted route').getAttribute('target')).toBe(
+      '_blank',
     );
   });
 });