+++ src/config/NonContentRoutes.jsx
@@ -1,3 +1,5 @@
+import config from '@plone/volto/registry';
+
 // Non Content Routes/Views
 // You can include either RegEx or a string representing the ending of the
 // nonContentRoute eg. '/add' will match '/foo/bar/add'
@@ -27,4 +29,5 @@ export const nonContentRoutes = [
   '/password-reset',
   '/create-translation',
   '/manage-translations',
+  ...(config.settings?.internalUrlBlacklist || []),
 ];