+++ src/config/index.js
@@ -138,6 +138,7 @@ let config = {
     showTags: true,
     controlPanelsIcons,
     showSelfRegistration: false,
+    internalUrlBlacklist: [], // URL to be considered as external
   },
   widgets: {
     ...widgetMapping,